fix: resolve 25 CodeQL scanning alerts in V2 code
- Add Severity comparison operators (eliminates 4 enum-undefined-behavior)
- Break path temporary chains into named variables (eliminates 15 use-after-expired-lifetime)
- Tighten bounds checks in exploit/security checks (eliminates 5 icc-buffer-overflow)
- Simplify enum comparisons (Command, icColorSpaceSignature)

+++ iccanalyzer-lite/icctest/cli/src/LinuxSandbox.cpp
@@ -34,7 +34,7 @@ namespace {
 Severity maxFindingSeverity(const std::vector<Finding>& findings) {
     Severity worst = Severity::INFO;
     for (const auto& finding : findings) {
-        if (static_cast<uint8_t>(finding.level) > static_cast<uint8_t>(worst)) {
+        if (finding.level > worst) {
             worst = finding.level;
         }
     }
